Move download types into shared project
