+++ src/shared/src/api/download.ts
@@ -1,0 +1,17 @@
+export type {
+  ByteRange,
+  Chunk,
+  ChunkProgress,
+  Chunker,
+  DownloadProgress,
+  Progress,
+  ProgressCallback,
+  ResolvedEndpoint,
+  ResolvedResource,
+  Resolver,
+  RetryContext,
+  RetryStrategy,
+  RetryVerdict,
+} from "../types/download";
+
+export { staticChunker, noRetry } from "../types/download";
