+++ src/shared/src/chunking.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from "vitest";
 
-import { type Chunk, staticChunker } from "./chunking";
+import { type Chunk, staticChunker } from "./types/download";
 
 function createChunks(size: number, numChunks: number): Chunk[] {
   return staticChunker(numChunks, 1)(size);
@@ -27,25 +27,25 @@ describe("staticChunker", () => {
   describe("coverage", () => {
     it("first chunk starts at byte 0", () => {
       const [first] = createChunks(100, 4);
-      expect(first.range.start).toBe(0);
+      expect(first!.range.start).toBe(0);
     });
 
     it("last chunk ends at the final byte", () => {
       const chunks = createChunks(100, 4);
-      expect(chunks.at(-1).range.end).toBe(99);
+      expect(chunks.at(-1)!.range.end).toBe(99);
     });
 
     it("covers every byte with no gaps", () => {
       const chunks = createChunks(100, 4);
       for (let i = 1; i < chunks.length; i++) {
-        expect(chunks[i].range.start).toBe(chunks[i - 1].range.end + 1);
+        expect(chunks[i]!.range.start).toBe(chunks[i - 1]!.range.end + 1);
       }
     });
 
     it("covers every byte with no gaps for an odd-sized file", () => {
       const chunks = createChunks(101, 4);
       for (let i = 1; i < chunks.length; i++) {
-        expect(chunks[i].range.start).toBe(chunks[i - 1].range.end + 1);
+        expect(chunks[i]!.range.start).toBe(chunks[i - 1]!.range.end + 1);
       }
     });
 
@@ -64,7 +64,7 @@ describe("staticChunker", () => {
     it("produces non-overlapping chunks", () => {
       const chunks = createChunks(100, 4);
       for (let i = 1; i < chunks.length; i++) {
-        expect(chunks[i].range.start).toBeGreaterThan(chunks[i - 1].range.end);
+        expect(chunks[i]!.range.start).toBeGreaterThan(chunks[i - 1]!.range.end);
       }
     });
 
@@ -100,8 +100,8 @@ describe("staticChunker", () => {
   describe("edge cases", () => {
     it("handles a single byte file", () => {
       const chunks = createChunks(1, 1);
-      expect(chunks[0].range.start).toBe(0);
-      expect(chunks[0].range.end).toBe(0);
+      expect(chunks[0]!.range.start).toBe(0);
+      expect(chunks[0]!.range.end).toBe(0);
     });
 
     it("throws when numChunks exceeds file size", () => {
@@ -110,18 +110,18 @@ describe("staticChunker", () => {
 
     it("handles a file that divides evenly into chunks", () => {
       const chunks = createChunks(100, 4);
-      expect(chunks[0].range).toMatchObject({ start: 0, end: 24 });
-      expect(chunks[1].range).toMatchObject({ start: 25, end: 49 });
-      expect(chunks[2].range).toMatchObject({ start: 50, end: 74 });
-      expect(chunks[3].range).toMatchObject({ start: 75, end: 99 });
+      expect(chunks[0]!.range).toMatchObject({ start: 0, end: 24 });
+      expect(chunks[1]!.range).toMatchObject({ start: 25, end: 49 });
+      expect(chunks[2]!.range).toMatchObject({ start: 50, end: 74 });
+      expect(chunks[3]!.range).toMatchObject({ start: 75, end: 99 });
     });
 
     it("handles a file that does not divide evenly into chunks", () => {
       const chunks = createChunks(101, 4);
-      expect(chunks[0].range).toMatchObject({ start: 0, end: 25 });
-      expect(chunks[1].range).toMatchObject({ start: 26, end: 51 });
-      expect(chunks[2].range).toMatchObject({ start: 52, end: 77 });
-      expect(chunks[3].range).toMatchObject({ start: 78, end: 100 });
+      expect(chunks[0]!.range).toMatchObject({ start: 0, end: 25 });
+      expect(chunks[1]!.range).toMatchObject({ start: 26, end: 51 });
+      expect(chunks[2]!.range).toMatchObject({ start: 52, end: 77 });
+      expect(chunks[3]!.range).toMatchObject({ start: 78, end: 100 });
     });
   });
 });
